ref(*): Adds support for arbitrary binary data
Per the [spec](https://github.com/rescrv/libmacaroons/blob/master/doc/format.txt):

> All fields other than the version and location fields may contain arbitrary
> binary data, though per-service conventions are free to impose stricter
> requirements - these are outside the scope of this document.

The current API only supported data that could be parsed as a string. This adds a
new `ByteString` type that handles the underlying base64 decoding and encoding as
part of its deserializing and serializing. There are also changes to the documentation.