ref(validator): Refactors validation logic
The validation logic was spread across various parts of the logic. This is a full
refactor of the validation logic that condenses it to a single reusable validator
for multiple macaroons. It also removes the need for "resetting" the validator
and switches out the current method of using a list of names into a set to track
usage of various discharge macaroons