ref(crypto): Introduces new MacaroonKey type
MacaroonKey is meant to be a vanity type for [u8; 32] with some additional
methods and conversions added to it. As part of this we now export the crypto module
to expose all of the MacaroonKey impls.

NOTE: I had to refactor some of the verifier tests to just construct the macaroon in
memory. I will be refactoring the verifier tests in the next commit