public API: flatten exports into simple namespace
This removes some public modules, like 'error', 'verify', 'crypto', and
'serialize'. Instead, simply export all the types (structs, enums, etc)
from the top level of the crate. Modules can be helpful for structuring
large APIs, but I think the API of this crate is small enough that they
can all be directly exported.

One concrete change in this commit is that the following functions in
the `macaroon::crypto` module would no longer be public (which I think
is desirable, they should not be part of the API surface):

- `decrypt_key`
- `encrypt_key`
- `hmac`
- `hmac2`