refactor top-level (de)serialization method behavior; make base64 flexible
The previous behavior matched libmacaroons upstream C API behavior, but
was confusing and inconsistent.

With these changes, both V1 and V2 tokens are assumed to be
base64-encoded by default, both when serializing and deserializing,
while V2JSON tokens are assumed to be simple JSON.

The difference between V1 and V2 tokens is still inferred based on the
first byte, after base64 decoding, which is always 0x02 for V2, or a hex
ASCII character for V1.

The top-level deserialize method is made generic via AsRef, which makes
it easier to pass in Strings, a common case for base64 tokens (or JSON).
The base64 decoding method is now generic over both URL-safe and
non-URL-safe encodings, as is common with other implementations and
mentioned in the libmacaroons docs.